# examples/path-conflict/openapi.original.yaml
openapi: 3.0.3
info:
  title: School
  version: "1"
security:
  - bearer: []
paths:
  /book:
    post:
      operationId: createBook
      tags: [Book]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Book" }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Book" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /book/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, minimum: 1 } }
    get:
      operationId: getBookById
      tags: [Book]
      responses:
        "200":
          description: the book
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Book" }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    put:
      operationId: updateBook
      tags: [Book]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Book" }
      responses:
        "200":
          description: updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Book" }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    delete:
      operationId: deleteBook
      tags: [Book]
      responses:
        "204": { description: deleted }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /book/{class}:
    parameters:
      - { name: class, in: path, required: true, schema: { type: string, pattern: "^[A-Z]$" } }
    get:
      operationId: getBookByClass
      tags: [Book]
      responses:
        "200":
          description: the books of the class
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Book" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    put:
      operationId: updateBookClass
      tags: [Book]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClassChange" }
      responses:
        "200":
          description: the books, moved to the new class
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Book" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    delete:
      operationId: deleteBookClass
      tags: [Book]
      responses:
        "204": { description: all books of the class deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /student:
    post:
      operationId: createStudent
      tags: [Student]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Student" }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Student" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /student/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, minimum: 1 } }
    get:
      operationId: getStudentById
      tags: [Student]
      responses:
        "200":
          description: the student
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Student" }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    put:
      operationId: updateStudent
      tags: [Student]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Student" }
      responses:
        "200":
          description: updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Student" }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    delete:
      operationId: deleteStudent
      tags: [Student]
      responses:
        "204": { description: deleted }
        "404": { $ref: "#/components/responses/NotFound" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /student/{class}:
    parameters:
      - { name: class, in: path, required: true, schema: { type: string, pattern: "^[A-Z]$" } }
    get:
      operationId: getStudentByClass
      tags: [Student]
      responses:
        "200":
          description: the students of the class
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Student" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    put:
      operationId: updateStudentClass
      tags: [Student]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ClassChange" }
      responses:
        "200":
          description: the students, moved to the new class
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Student" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    delete:
      operationId: deleteStudentClass
      tags: [Student]
      responses:
        "204": { description: all students of the class deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer }
  responses:
    BadRequest:
      description: invalid input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: no or invalid token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Book:
      type: object
      required: [Title, Class]
      properties:
        Id: { type: integer, readOnly: true }
        Title: { type: string, minLength: 1, maxLength: 60 }
        Class: { type: string, pattern: "^[A-Z]$" }
    Student:
      type: object
      required: [Name, Class]
      properties:
        Id: { type: integer, readOnly: true }
        Name: { type: string, minLength: 1, maxLength: 60 }
        Class: { type: string, pattern: "^[A-Z]$" }
    ClassChange:
      type: object
      required: [Class]
      properties:
        Class: { type: string, pattern: "^[A-Z]$" }
    Problem:
      type: object
      required: [Message]
      properties:
        Message: { type: string }